Troubleshooting bugs
Experimenting with various changes, hoping to solve bug: printtime elapsed stops counting when paused (as it should) but does not restart when Resume is selected.
The Dsuuuu design uses ust print_status screen, with buttons that appear and disappear, according to whether paused or resumed.  My design uses two different screens.  Suspect there is now code designed for the original approach that is not working with the new approach.   Need to find and fix it.  Still looking...

Message (M117) text is formatted by t5uid1. Was not justifying correctly in Text or Rolling Text widgets.  Have achieved the display I want, but have not figured out why I had to format 32 bits long & center-justified, for a left-justified display of 20 characters...