Release 1.2.4 now includes pre-configured printer.cfg files for all CR6 motherboards
Added a zipfile containing annotated Klipper configuration files for each type of CR6 motherboard.  Users can now perform a full clean-install of Mainsail OS to an SD card,  upload all of the config and cfg files as-is, and then tailor the printer.cfg file to define their system.

A DRAFT installation guide has been provided in the docs folder to help users perform a clean install.  Feedback is welcome. Expect the documentation to improve in the next release.