Update Release Notes_v1.2.5.txt
I am updating this component to support the release of v0.3.7 of the DWIN_SET component, which has been enhanced to support optimizing the bed mesh using the RunABL screen when tweaking bed screws and silicone spacers.

I encountered one occasion when the HEIGHTMAP function displayed the entire mesh below the grid.  I suspected (but did not prove) that it had appended the new probe map to the previous bed mesh and wanted to eliminate that possibility.