Updated the MACHINE config files for the 4.5.2 and BTT motherboards
I believe the only necessary differences between 4.5.3 and 4.5.2 boards are the pin numbers for the two heaters and the part cooling fan.  Those are therefore the only changes I made.
I believe the differences between the 4.5.3 and the BTT SKR CR6 board are the TMC2209 settngs + all of the motor microsteps (64 vs 16)v+ all of the pin numbers.
I changed all of those.