Fix Issue#37
SAVE_CONFIG was failing, if the brightness level was modified with the SetUp sceen control.
Solution instead relies on t5uid1.py to default brightness to 100 if no other value has been stored on the display. Otherwise, the display will remain at the last display brightness level set, even after power-cycling or RESTART or FIRMWARE RESTART of the printer.