Modified x, y and z position displays to show "0.0" in RED if NOT HOMED
As with CF6.1, the X, Y and Z position displays now turn RED and display 0.0, if the axes are not homed.
Same thing will happen if the motors are disabled or if they time-out.
The obective is to alert users to situations where homing is required before axes can be moved.

Requires DWIN_SET 0.4.0, in which the SP values for each of the 3 variables are now defined.
(X = 0x5000, Y=0x5010 Z=0x5020)