Update README.md with Slicer Setting Guidelines
Users of OrcaSlicer alerted me to the fact that some macros use gcodes variable names used by Cura but that creates incompatibility problems with OrcaSlicer and other slicers) 
I have experimented a little with OrcaSlicer 2.0.0, and made some adustments, accordingly.  (I do not plan to experiment with other slicers, though.  I you use something else, you are welcome to raise an Issue or Discussion, to request additional guidelines be added to this Readme.

Signed-off-by: Thinkersbluff <36551518+Thinkersbluff@users.noreply.github.com>