Restore line 928 of t5uid1.py to original logic
Rev 1.3.2 replaced the "If probe is not None..." logic test with a the value "False", to provide a quick fix to the bugs introduced by a Klipper mod in June 2024, while I researched the correct "fix".
This mod now implements the "correct" fix, by updating the address of the multi_probe_pending property, which changed when probe.py was refactored.