Minor bug fixes and enhancements
Updated the Information page to fix issue#38.
Added a variable to vars_out.cfg, to brand the Klipper Component with its version number.  This value is read and displayed on the Information Page at DWIN_SET version 0.4.4 and subs.
Modified the routine that updates the print time parameters, to delay the print time remaining "countdown" until the printer starts drawing the purge line, since the pre-heating and initial probing times are not included in the slicer's estimate of the print duration.