Refine the Slicer-Estimated Print Time Remaining to wait for the purge line to begin
Previously, the countdown timer started when the print button was pressed, which meant that the time spent warming-up and homing the printer was being subtracted from the estimated print time.  For very short prints, this caused the time to count down to zero minutes before the print actually finished.
These mods pause the countdown, measure the startup time, and then add that back into the total time remaining, when the purge line begins.